refactor: remove old chunkmap
